+++ json_serializable/test/generic_files/generic_class.g.dart
@@ -11,19 +11,21 @@ GenericClass<T, S> _$GenericClassFromJson<T extends num, S>(
   return GenericClass<T, S>()
     ..fieldObject = json['fieldObject'] == null
         ? null
-        : _dataFromJson(json['fieldObject'] as Map<String, dynamic>)
+        : GenericClass._dataFromJson(
+            json['fieldObject'] as Map<String, dynamic>)
     ..fieldDynamic = json['fieldDynamic'] == null
         ? null
-        : _dataFromJson(json['fieldDynamic'] as Map<String, dynamic>)
+        : GenericClass._dataFromJson(
+            json['fieldDynamic'] as Map<String, dynamic>)
     ..fieldInt = json['fieldInt'] == null
         ? null
-        : _dataFromJson(json['fieldInt'] as Map<String, dynamic>)
+        : GenericClass._dataFromJson(json['fieldInt'] as Map<String, dynamic>)
     ..fieldT = json['fieldT'] == null
         ? null
-        : _dataFromJson(json['fieldT'] as Map<String, dynamic>)
+        : GenericClass._dataFromJson(json['fieldT'] as Map<String, dynamic>)
     ..fieldS = json['fieldS'] == null
         ? null
-        : _dataFromJson(json['fieldS'] as Map<String, dynamic>);
+        : GenericClass._dataFromJson(json['fieldS'] as Map<String, dynamic>);
 }
 
 Map<String, dynamic> _$GenericClassToJson<T extends num, S>(
@@ -31,12 +33,17 @@ Map<String, dynamic> _$GenericClassToJson<T extends num, S>(
     <String, dynamic>{
       'fieldObject': instance.fieldObject == null
           ? null
-          : _dataToJson(instance.fieldObject),
+          : GenericClass._dataToJson(instance.fieldObject),
       'fieldDynamic': instance.fieldDynamic == null
           ? null
-          : _dataToJson(instance.fieldDynamic),
-      'fieldInt':
-          instance.fieldInt == null ? null : _dataToJson(instance.fieldInt),
-      'fieldT': instance.fieldT == null ? null : _dataToJson(instance.fieldT),
-      'fieldS': instance.fieldS == null ? null : _dataToJson(instance.fieldS)
+          : GenericClass._dataToJson(instance.fieldDynamic),
+      'fieldInt': instance.fieldInt == null
+          ? null
+          : GenericClass._dataToJson(instance.fieldInt),
+      'fieldT': instance.fieldT == null
+          ? null
+          : GenericClass._dataToJson(instance.fieldT),
+      'fieldS': instance.fieldS == null
+          ? null
+          : GenericClass._dataToJson(instance.fieldS)
     };